feat: moved postTransformPublicPath tests to own file
